reshuffle code out of the node
